requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
nltk
langdetect
deep-translator
newspaper3k
matplotlib
seaborn
